+++ api/endpoints/swipes.py
@@ -63,8 +63,11 @@ def volunteer_swipe(request, taskid, pinid, match):
         )
 
         if match == True: # This means both parties swiped right.
-            # Implement this!!!
-            return HttpResponse("Not implemented", status=501)
+            new_conversation(pinid, volunteerid, taskid)
+            return HttpResponse(
+                "Match detected",
+                status=200
+            )
 
     except ValidationError:
         return HttpResponse(
